pyproject.toml
[tool.poetry]
name = "auto-toot-delete"
version = "0.1.0"
description = ""
authors = ["fad0805 <s222y@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"Mastodon.py" = "^1.8.1"
parse = "^1.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
